Add Figma Make rules for getadb
