+++ client/www/app/getadb/route.ts
@@ -1,10 +1,20 @@
-import { agentMarkdown } from './markdown';
+import { getServerConfig } from '@/lib/config';
+import generateMarkdown from './generateMarkdown';
 
 export const dynamic = 'force-dynamic';
 export const runtime = 'nodejs';
 
 export async function GET(request: Request) {
-  const markdown = await agentMarkdown(request);
+  const token = process.env.GET_A_DB_PERSONAL_ACCESS_TOKEN;
+  if (!token) {
+    throw new Error('GET_A_DB_PERSONAL_ACCESS_TOKEN is not set');
+  }
+  const title =
+    new URL(request.url).searchParams.get('title')?.trim() || DEFAULT_APP_TITLE;
+  const app = await createApp(token, title);
+
+  const markdown = await generateMarkdown(request, app);
+
   return new Response(markdown, {
     headers: {
       'Cache-Control': 'private, no-store, max-age=0',
@@ -12,4 +22,30 @@ export async function GET(request: Request) {
       'Content-Type': 'text/markdown; charset=utf-8',
     },
   });
+}
+
+const DEFAULT_APP_TITLE = 'Instant App';
+
+async function createApp(
+  token: string,
+  title: string,
+): Promise<{ id: string; adminToken: string }> {
+  const { apiURI } = await getServerConfig();
+  const res = await fetch(`${apiURI}/dash/apps/get_a_db`, {
+    method: 'POST',
+    headers: {
+      'Content-Type': 'application/json',
+      Authorization: `Bearer ${token}`,
+    },
+    body: JSON.stringify({ title }),
+  });
+  if (!res.ok) {
+    throw new Error(
+      `Failed to create get-a-db app: ${res.status} ${await res.text()}`,
+    );
+  }
+  const body = (await res.json()) as {
+    app: { id: string; 'admin-token': string };
+  };
+  return { id: body.app.id, adminToken: body.app['admin-token'] };
 }
